pyproject.toml
[project]
name = "ai-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.10.0",
    "google-search-results>=2.4.2",
    "jinja2>=3.1.6",
    "langchain>=0.3.27",
    "langchain-huggingface>=0.3.1",
    "langchainhub>=0.1.21",
    "numexpr>=2.11.0",
    "openai>=1.99.3",
    "pydantic-ai>=0.1.0",
    "sentencepiece>=0.2.1",
    "text-generation>=0.7.0",
    "transformers>=4.55.0",
]
